requirements.txt
# Streamlit e GUI
streamlit>=1.24

# Data processing
pandas>=1.5

# Estrazione testo da PDF/DOCX/TXT
pdfplumber>=0.10
python-docx>=0.8

# LLM client
openai<=0.27

# Modelli Transformer e NER
transformers>=4.30
torch>=2.0
httpx==0.27.2

# GLiNER (NER PII)
gliner>=0.1

# Dipendenza necessaria di pdfplumber/pdfminer
charset-normalizer<3.1.0
